requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.19.0
numpy>=1.26.0,<2.2.0
pillow==10.0.1
scikit-learn==1.3.0 